backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
supabase
pydantic-settings
websockets
python-dotenv
pytest
httpx
pypdf
openai
unstructured[pdf]
pdfminer.six
